Adding function to rebin a 2d array

+++ IceBRG_main/IceBRG_main/Eigen.hpp
@@ -69,7 +69,7 @@ auto end(Eigen::Array<Derived,Nrow,Ncol,Options,Maxrow,Maxcol> & vec) -> decltyp
 
 namespace IceBRG {
 
-template<typename T, int Nrow=Eigen::Dynamic, int Ncol=1>
+template<typename T, ssize_t Nrow=Eigen::Dynamic, ssize_t Ncol=1>
 using array_t = Eigen::Array<T,Nrow,Ncol>;
 
 typedef array_t<flt_t> flt_array_t;
